\name{Covariates}
\alias{Covariates}
\docType{data}
\title{
	Artificial data for Matrix eQTL sample code: Covariates.
}
\description{
	Artificial data set with 2 covariates across 15 samples. 
	Columns of the file must match to those of the expression and genotype data sets.
}
\format{
	\tabular{llll}{
		id     \tab Sam_01   \tab Sam_02 \tab \ldots \cr
		gender \tab 0        \tab 0      \tab \ldots \cr
		age    \tab 40        \tab 46    \tab \ldots \cr
		
	}
}
\references{
	The package website: \url{http://www.bios.unc.edu/research/genomic_software/Matrix_eQTL/}
}
\seealso{
	See \code{\link{Matrix_eQTL_engine}} for reference and sample code.
}
\keyword{datasets}
