\name{GE}
\alias{GE}
\docType{data}
\title{
	Artificial data for Matrix eQTL sample code: Gene expression.
}
\description{
	Artificial data set with expression of 10 genes across 15 samples. 
	Columns of the file must match to those of the genotype and covariates data sets.
}
\format{
	\tabular{llll}{
		geneid  \tab Sam_01   \tab Sam_02 \tab \ldots \cr
		Gene_01 \tab 4.91     \tab 4.63   \tab \ldots \cr
		Gene_02 \tab 13.78    \tab 13.14  \tab \ldots \cr
		\ldots  \tab \ldots   \tab \ldots \tab \ldots \cr
	}
}
\references{
	The package website: \url{http://www.bios.unc.edu/research/genomic_software/Matrix_eQTL/}
}
\seealso{
	See \code{\link{Matrix_eQTL_engine}} for reference and sample code.
}
\keyword{datasets}
