\name{modelANOVA}
\alias{modelANOVA}
\docType{data}
\title{
Constant for \code{\link{Matrix_eQTL_engine}}.
}
\description{
Use of the constant as a parameter for \code{\link{Matrix_eQTL_engine}} 
to indicates that the genotype should be treated as a categorical variable.
}
\usage{
modelANOVA
}