\name{modelLINEAR}
\alias{modelLINEAR}
\docType{data}
\title{
Constant for \code{\link{Matrix_eQTL_engine}}.
}
\description{
Use of the constant as a parameter for \code{\link{Matrix_eQTL_engine}} 
to indicates that the effect of genotype on expression is modeled to be linear.
}
\usage{
modelLINEAR
}