\name{SNP}
\alias{SNP}
\docType{data}
\title{
	Artificial data for Matrix eQTL sample code: Genotype.
}
\description{
	Artificial data set with genotype information for 15 markers across 15 samples.
	Columns of the file must match to those of the gene expression and covariates data sets.
}
\format{
	\tabular{llll}{
		snpnid  \tab Sam_01   \tab Sam_02 \tab \ldots \cr
		Snp_01  \tab 2        \tab 0      \tab \ldots \cr
		Snp_02  \tab 0        \tab 1      \tab \ldots \cr
		\ldots  \tab \ldots   \tab \ldots \tab \ldots \cr
	}
}
\references{
	The package website: \url{http://www.bios.unc.edu/research/genomic_software/Matrix_eQTL/}
}
\seealso{
	See \code{\link{Matrix_eQTL_engine}} for reference and sample code.
}
\keyword{datasets}
