\name{MatrixeQTL-package}
\alias{MatrixEQTL-package}
\alias{MatrixeQTL}
\alias{MatrixEQTL}
\docType{package}
\title{
Matrix eQTL: Ultra fast eQTL analysis via large matrix operations
}
\description{
Matrix eQTL is designed for fast eQTL analysis on large datasets.
Matrix eQTL can model and test for association between genotype and gene expression using both linear regression and ANOVA models. 
The models can include covariates to account for such factors as population structure, gender, and clinical variables. 
It also supports testing of heteroscedastic models and models with correlated errors.
}
\details{
\tabular{ll}{
Package: \tab MatrixeQTL\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-06-25\cr
License: \tab LGPL-3 \cr
LazyLoad: \tab yes\cr
Depends: \tab methods\cr
}
}
\author{
Andrey Shabalin \email{shabalin@email.unc.edu}

Maintainer: Andrey Shabalin <shabalin@email.unc.edu>
}
\references{
The software website:
\url{http://www.bios.unc.edu/research/genomic_software/Matrix_eQTL/}
}
\keyword{ package }
\keyword{ MatrixeQTL }
\keyword{ Matrix eQTL }
\keyword{ Andrey Shabalin }
\keyword{ Shabalin }
\seealso{
See \code{\link{Matrix_eQTL_engine}} for the description and sample code.
}